// src/types/enums.ts
/**
 * Keep in sync with backend
 */


export enum DietType {
    vegan = "vegan",
    vegetarian = "vegetarian",
    fish = "fish",
    seafood = "seafood",
    poultry = "poultry",
    beef = "beef",
    pork = "pork",
    other = "other"
}

export enum MealType {
    breakfast = "breakfast",
    lunch = "lunch",
    dinner = "dinner",
    snack = "snack",
    dessert = "dessert",
    drink = "drink",
    other = "other"
}

export enum RecipeTags {
    quick = "quick",
    easy = "easy",
    healthy = "healthy",
    cheap = "cheap",
    spicy = "spicy",
    sweet = "sweet",
    salty = "salty",
    savory = "savory",
    satisfying = "satisfying",
    filling = "filling",
    light = "light",
    heavy = "heavy"
}